# kisaan-backend-node/openapi.yaml
openapi: 3.0.3
info:
  title: KisaanCenter API
  version: 1.0.0
  description: Full API documentation for KisaanCenter backend
servers:
  - url: http://localhost:3000/api
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Product:
      type: object
      properties:
        id: { type: integer }
        name: { type: string }
        category_id: { type: integer }
        status: { type: string, enum: [active, inactive], default: active }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
      required: [name, category_id]
...
    Payment:
      type: object
      properties:
        id: { type: integer }
        transaction_id: { type: integer }
        payer_type: { type: string, enum: [BUYER, SHOP] }
        payee_type: { type: string, enum: [SHOP, FARMER] }
        amount: { type: number, minimum: 0 }
        status: { type: string, enum: [PENDING, PAID, FAILED], default: PENDING }
        payment_date: { type: string, format: date-time, nullable: true }
        method: { type: string, enum: [CASH, BANK, UPI, OTHER], default: CASH }
        notes: { type: string, nullable: true }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
      required: [transaction_id, payer_type, payee_type, amount]
    PaymentCreate:
      type: object
      properties:
        transaction_id: { type: integer }
        payer_type: { type: string, enum: [BUYER, SHOP] }
        payee_type: { type: string, enum: [SHOP, FARMER] }
        amount: { type: number, minimum: 0 }
        method: { type: string, enum: [CASH, BANK, UPI, OTHER], default: CASH }
        notes: { type: string, nullable: true }
        payment_date: { type: string, format: date-time, nullable: true }
      required: [transaction_id, payer_type, payee_type, amount]
    Settlement:
      type: object
      properties:
        id: { type: integer }
        shop_id: { type: integer }
        user_id: { type: integer }
        amount: { type: number }
        reason: { type: string, enum: [overpayment, underpayment, adjustment] }
        status: { type: string, enum: [pending, settled], default: pending }
        settlement_date: { type: string, format: date-time, nullable: true }
        notes: { type: string, nullable: true }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
      required: [shop_id, user_id, amount, reason]
    SettlementCreate:
      type: object
      properties:
        shop_id: { type: integer }
        user_id: { type: integer }
        amount: { type: number }
        reason: { type: string, enum: [overpayment, underpayment, adjustment] }
        notes: { type: string, nullable: true }
      required: [shop_id, user_id, amount, reason]
...
    PaginatedResponse:
      type: object
      properties:
        data: { type: array }
        total: { type: integer }
        page: { type: integer }
        limit: { type: integer }
        totalPages: { type: integer }
    ApiResponse:
      type: object
      properties:
        success: { type: boolean }
        message: { type: string }
        data: { }
    ErrorResponse:
      type: object
      properties:
        success: { type: boolean, default: false }
        message: { type: string }
        error: { type: string }
...
paths:
...
  /payments:
    post:
      summary: Create payment
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/PaymentCreate' }
      responses:
        '201':
          description: Payment created
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data: { $ref: '#/components/schemas/Payment' }
        '400':
          description: Bad request
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
    get:
      summary: Get all payments
      parameters:
        - in: query
          name: transaction_id
          schema: { type: integer }
        - in: query
          name: status
          schema: { type: string, enum: [PENDING, PAID, FAILED] }
        - in: query
          name: payer_type
          schema: { type: string, enum: [BUYER, SHOP] }
        - in: query
          name: payee_type
          schema: { type: string, enum: [SHOP, FARMER] }
        - in: query
          name: page
          schema: { type: integer, minimum: 1, default: 1 }
        - in: query
          name: limit
          schema: { type: integer, minimum: 1, maximum: 100, default: 10 }
      responses:
        '200':
          description: List of payments
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/PaginatedResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items: { $ref: '#/components/schemas/Payment' }
  /payments/{id}:
    get:
      summary: Get payment by ID
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        '200':
          description: Payment found
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data: { $ref: '#/components/schemas/Payment' }
        '404':
          description: Payment not found
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
  /payments/{id}/status:
    put:
      summary: Update payment status
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                status: { type: string, enum: [PENDING, PAID, FAILED] }
                payment_date: { type: string, format: date-time }
                notes: { type: string }
      responses:
        '200':
          description: Payment status updated
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data: { $ref: '#/components/schemas/Payment' }
  /payments/transaction/{transactionId}:
    get:
      summary: Get payments by transaction
      parameters:
        - in: path
          name: transactionId
          required: true
          schema: { type: integer }
      responses:
        '200':
          description: Transaction payments
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items: { $ref: '#/components/schemas/Payment' }
...
  /payments/farmers/{farmerId}:
    get:
      summary: Get payments to farmer
      parameters:
        - in: path
          name: farmerId
          required: true
          schema: { type: integer }
      responses:
        '200':
          description: Farmer payments
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/PaginatedResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items: { $ref: '#/components/schemas/Payment' }
  /payments/buyers/{buyerId}:
    get:
      summary: Get payments by buyer
      parameters:
        - in: path
          name: buyerId
          required: true
          schema: { type: integer }
      responses:
        '200':
          description: Buyer payments
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/PaginatedResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items: { $ref: '#/components/schemas/Payment' }
  /settlements:
    post:
      summary: Create settlement
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/SettlementCreate' }
      responses:
        '201':
          description: Settlement created
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data: { $ref: '#/components/schemas/Settlement' }
        '400':
          description: Bad request
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
    get:
      summary: Get all settlements
      parameters:
        - in: query
          name: shop_id
          schema: { type: integer }
        - in: query
          name: user_id
          schema: { type: integer }
        - in: query
          name: status
          schema: { type: string, enum: [pending, settled] }
        - in: query
          name: reason
          schema: { type: string, enum: [overpayment, underpayment, adjustment] }
        - in: query
          name: page
          schema: { type: integer, minimum: 1, default: 1 }
        - in: query
          name: limit
          schema: { type: integer, minimum: 1, maximum: 100, default: 10 }
      responses:
        '200':
          description: List of settlements
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/PaginatedResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items: { $ref: '#/components/schemas/Settlement' }
  /settlements/{id}:
    get:
      summary: Get settlement by ID
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        '200':
          description: Settlement found
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data: { $ref: '#/components/schemas/Settlement' }
        '404':
          description: Settlement not found
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
    put:
      summary: Update settlement
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                status: { type: string, enum: [pending, settled] }
                settlement_date: { type: string, format: date-time }
                notes: { type: string }
      responses:
        '200':
          description: Settlement updated
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data: { $ref: '#/components/schemas/Settlement' }
        '404':
          description: Settlement not found
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
...
  /settlements/settle/{settlement_id}:
    post:
      summary: Settle amount
      parameters:
        - in: path
          name: settlement_id
          required: true
          schema: { type: integer }
      responses:
        '200':
          description: Amount settled
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiResponse' }
  /settlements/expense:
    post:
      summary: Create expense
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                shop_id: { type: integer }
                amount: { type: number, minimum: 0 }
                description: { type: string }
              required: [shop_id, amount, description]
      responses:
        '201':
          description: Expense created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiResponse' }
...
  /shops/{id}/products:
    get:
      summary: Get products for shop
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        '200':
          description: Shop products
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items: { $ref: '#/components/schemas/Product' }
...